Add database file generation and improve error handling
This update introduces the generation of a database file if it doesn't already exist for the TradesPlugin. Also, there are changes to improve the exception handling in the code: exceptions thrown while adding a trade to the database now return a message to notify users of the specific error. Finally, small alterations were made to the text strings in the CommandTrade and CommandDeleteTrade files.